Binomial data distribution

Implements the normal approximation for a test on rates. The reponse rate in the control group, rC, has to be specified by rate_control. The null hypothesis is: rE ≤ rC, where rE denotes the response rate in the invervention group. It is tested against the alternative rE > rC. The test statistic is given as X1 = √n (rE - rC) / √(2 r0 (1-r0)), where r0 denotes the mean between rE and rC in the two-armed case, and rE in the one-armed case.#' All priors have to be defined for the rate difference rE - rC.

Binomial(rate_control, two_armed = TRUE)

# S4 method for Binomial
quantile(x, probs, n, theta, ...)

# S4 method for Binomial,numeric
simulate(object, nsim, n, theta, seed = NULL, ...)

Arguments

rate_control
assumed response rate in control group

two_armed
logical indicating if a two-armed trial is regarded

x
outcome

probs
vector of probabilities

n
sample size

theta
distribution parameter

...
further optional arguments

object
object of class Binomial

nsim
number of simulation runs

seed
random seed

Details

Note that simulate for class Binomial simulates the normal approximation of the test statistic.

Slots

rate_control

cf. parameter 'rate_control'

See also

Examples

datadist <- Binomial(rate_control = 0.2, two_armed = FALSE)