Gaussian-quadrature based implementation of Design

Implements a class for two-stage designs where n2 and c2 are approximated by linear interpolation of pivot points and integration is handled via Gaussian quadrature rules. The pivots are chosen to be the evaluation points of the selected Gaussian quadrature rule and guarantee a stable evaluation during optimization.

[TODO: implement a constructor taking in n2 and c2 functions?]

GQDesign(n1, c1f, c1e, n2_pivots, c2_pivots, order)

# S4 method for GQDesign
update(object, params, ...)

# S4 method for GQDesign
n1(d, ...)

# S4 method for GQDesign
get_knots(d, ...)

# S4 method for GQDesign,numeric
n2(d, x1, ...)

# S4 method for GQDesign,numeric
c2(d, x1, ...)

# S4 method for GQDesign
as.numeric(x, ...)

Arguments

n1
stage-one sample size

c1f
early stopping for futility boundary

c1e
early stopping for efficacy boundary

n2_pivots
vector of length order giving the values of n2 at the pivot points of the Gaussian quadrature rule [TODO: these are not available during construction]

c2_pivots
vector of length order giving the values of c2 at the pivot points of the Gaussian quadrature rule [TODO: these are not available during construction]

order
order (i.e. number of pivot points in the interior of [c1f, c1e]) of the Gaussian quadrature rule to use for integration

object
GQDesign to update with given parameter values.

d
object of class GQDesign

x1
stage-one outcome

x
GQDesign

Methods (by generic)

update: return a new design object with parameters given by a single numeric vector of length 2*order + 3, inverse of as.numeric(design).

n1: stage-one sample size

get_knots: get the pivots points (knots) of the Gaussian quadrature rule.

n2: stage-two sample size by linear interpolation of n2_pivots

c2: stage-two critical value by linear interpolation of n2_pivots

as.numeric: convert GQDesign to vector of tunable parameters via c(x@n1, x@c1f, x@c1e, x@n2_pivots, x@c2_pivots); inverse of update();

Slots

rule

data.frame, the integration rule, result of a call to gaussquad::legendre.quadrature.rules(order)[[order]]) during construction