Point-mass prior

PointMassPrior is a generic class for representing a univariate prior over a discrete set of points with positive probability mass. All methods for prior distributions [TODO link] are specialized to the discrete case.

PointMassPrior(theta, mass)

# S4 method for PointMassPrior
bounds(dist, ...)

# S4 method for PointMassPrior,`function`
expectation(dist, f, ...)

# S4 method for PointMassPrior,numeric
condition(dist, interval, ...)

# S4 method for DataDistribution,PointMassPrior,numeric
predictive_pdf(dist,
  prior, x1, n1, ...)

# S4 method for DataDistribution,PointMassPrior,numeric
predictive_cdf(dist,
  prior, x1, n1, ...)

# S4 method for DataDistribution,PointMassPrior,numeric
posterior(dist, prior,
  x1, n1, ...)

Arguments

theta
cf. slot 'theta'

mass
cf. slot 'mass'

dist
data distribution or prior distribution (depending on context) [TODO: make this clearer by renaming arguments in generics!]

...
further optional arguments

f
univariate function of parameter to get expected value of

interval
numeric vector of length two specifying the interval to condition on.

prior
prior distribution

x1
stage one outcome (note that n1 is available from design)

n1
stage-one sample size

Slots

theta

numeric vector of pivot points (parameter values with positive prior mass)

mass

numeric vector of the same length as theta - corresponding probability masses (must sum to one!)