Abstract univariate prior on model parameter

Prior is an abstract class specifying the interface to be implemented by an prior class. Implementations for PointMassPrior and ContinuousPrior are available.

bounds() returns the range of the support of a prior [TODO: rename support()?]

expectation() evaluates the expected value of a function f depending on only the parameter value with respect to the prior distribution dist.

condition() returns a new object of class prior resulting from conditioning on the given interval.

predictive_pdf() evaluates the predictive pdf of the model specified by a data distribution dist and prior at he given stage-one outcome (returns PMF for discrete priors).

predictive_cdf() evaluates the predictive cdf of the model specified by a data distribution dist and prior at he given stage-one outcome.

posterior() returns a new prior distribution [TODO: okay, maybe rethink naming?] resulting from conditioning on a given stage-one outcome.

bounds(dist, ...)

expectation(dist, f, ...)

condition(dist, interval, ...)

predictive_pdf(dist, prior, x1, n1, ...)

predictive_cdf(dist, prior, x1, n1, ...)

posterior(dist, prior, x1, n1, ...)

Arguments

dist
a distribution object of class Prior or DataDistribution

...
further optional parameters

f
a function of the single model parameter, must be vectorized, i.e. accept a vector of paramter values and return the vector of function values.

interval
length two numeric vector giving the parameter interval to condition on

prior
a distribution object of class Prior

x1
stage-one outcome

n1
stage-one sample size