Score Composition

composite defines new composite scores by point-wise evaluation of scores in any valid numerical expression.

composite(expr, label = NA_character_)

# S4 method for CompositeScore,TwoStageDesign
evaluate(s, design, ...)

Arguments

expr
Expression (in curly brackets); must contain at least one score variable; if multiple scores are used, they must either all be conditional or unconditional. Currently, no non-score variables are supported

label
object label (string)

s
object of class CompositeScore

design
object

...
further optional arguments

Value

an object of class CompositeConditionalScore or CompositeUnconditionalScore depending on the class of the scores used in expr

See also

Examples

ess   <- ExpectedSampleSize(Normal(), PointMassPrior(.4, 1))
power <- Power(Normal(), PointMassPrior(.4, 1))

# linear combination:
composite({ess - 50*power})

#> E[E[n(x1)]]<Normal<two-armed>;PointMass<0.40>>  - 50 * E[Pr[x2>=c2(x1)]]<Normal<two-armed>;PointMass<0.40>>  

# control flow (e.g. for and while loops)
composite({
  res <- 0
  for (i in 1:3) {
     res <- res + ess
  }
  res
})

#> res <- 0; for (i in 1:3) {
#>     res <- res + E[E[n(x1)]]<Normal<two-armed>;PointMass<0.40>> 
#> }; res 

# functional composition
composite({log(ess)})

#> log(E[E[n(x1)]]<Normal<two-armed>;PointMass<0.40>> ) 

cp <- ConditionalPower(Normal(), PointMassPrior(.4, 1))
composite({3*cp})

#> 3 * Pr[x2>=c2(x1)|x1]<Normal<two-armed>;PointMass<0.40>>