Group-sequential design

The class GroupSequentialDesign allows the implementation and usage of group-sequential designs. These are two-stage designs with a constant stage-two sample size function. Therefore, the length of n2_pivots is required to be 1 in this class. GroupSequentialDesign is a subclass of TwoStageDesign.

If order is not defined, the length of the parameter c2_pivots defines the order of the Gaussian quadrature integration rule.

Only states the class itself.

GroupSequentialDesign(n1, c1f, c1e, n2_pivots, c2_pivots, order = NULL,
  ...)

# S4 method for GroupSequentialDesign,numeric
n2(d, x1, round = TRUE, ...)

# S4 method for GroupSequentialDesign
TwoStageDesign(d, tunable, ...)

# S4 method for GroupSequentialDesign
show(object)

# S4 method for SmoothnessN2,GroupSequentialDesign
evaluate(s, design, ...)

Arguments

n1
cf. slot

c1f
cf. slot

c1e
cf. slot

n2_pivots
stage-two sample size

c2_pivots
stage-two rejection boundary

order
order of Gaussian quadrature

...
further optional arguments

d
object of class GSDesign

x1
stage-one outcome

round
logical, should integer sample size or real sample size be returned?

tunable
c.f. slot

object
design to show

s
an object of class SmoothnessN2

design
an object of class GroupSequentialDesign

Value

an object of class GroupSequentialDesign

Slots

n1

stage-one sample size

c1f

early stopping for futility boundary

c1e

early stopping for efficacy boundary

n2_pivots

stage-two sample size upon continuation

c2_pivots

vector of length order giving the values of c2 at the pivot points of the numeric integration rule [TODO: these are not available during construction]

x1_norm_pivots

normalized pivots for integration rule (in [-1, 1])

weights

weights of conditional score values at x1_norm_pivots for approximating the integral over x1.

tunable

named logical vector indicating whether corresponding slot is considered a tunable parameter