One-stage design

OneStageDesign allows the usage of a single-stage design. In this case without a second stage, there only exist the first-stage sample size n1 and the first-stage rejection boundary c1f. No other parameters have to be set. OneStageDesign is a subclass of TwoStageDesign.

OneStageDesign(n, c)

# S4 method for OneStageDesign
update(object, params, ...)

# S4 method for OneStageDesign,numeric
n2(d, x1, ...)

# S4 method for OneStageDesign,numeric
c2(d, x1, ...)

# S4 method for OneStageDesign
TwoStageDesign(d, ...)

# S4 method for OneStageDesign
plot(x, y, ...)

# S4 method for SmoothnessN2,OneStageDesign
evaluate(s, design, ...)

Arguments

n
sample size (stage-one sample size)

c
rejection boundary (c = c1f = c1e)

object
object to update

params
vector of design parameters (must be in same order as returned by as.numeric(OneStageDesign))

...
further optional arguments

d
object of class OneStageDesign

x1
stage-one outcome

x
not used

y
not used

s
an object of class SmoothnessN2

design
an object of class OneStageDesign

Slots

n1

sample size

c1f

rejection boundary

c1e

rejection boundary

n2_pivots

constant 0

c2_pivots

+/- Inf

x1_norm_pivots

is ignored for OneStageDesign

weights

is ignored for OneStageDesign

tunable

is ignored for OneStageDesign