Adaptive Optimal Two-Stage Designs

The adoptr package provides functionality to explore custom optimal two-stage designs for one- or two-arm superiority tests. For more details on the theoretical background see doi:10.1002/sim.8291 and doi:10.18637/jss.v098.i09. adoptr makes heavy use of the S4 class system. A good place to start learning about it can be found here.

Quickstart

For a sample workflow and a quick demo of the capabilities, see here.

A more detailed description of the background and the usage of adoptr can be found here or here doi:10.18637/jss.v098.i09 .

A variety of examples is presented in the validation report hosted here.

Designs

adoptr currently supports TwoStageDesign, GroupSequentialDesign, and OneStageDesign.

Data distributions

The implemented data distributions are Normal, Binomial, Student, Survival, ChiSquared (including Pearson2xK and ZSquared) and ANOVA.

Priors

Both ContinuousPrior and PointMassPrior are supported for the single parameter of a DataDistribution.

Scores

See Scores for information on the basic system of representing scores. Available scores are ConditionalPower, ConditionalSampleSize, Power, and ExpectedSampleSize.

See also

Author

Maintainer: Maximilian Pilz maximilian.pilz@itwm.fraunhofer.de (ORCID)

Authors:

Kevin Kunzmann kevin.kunzmann@boehringer-ingelheim.com (ORCID) [copyright holder]

Jan Meis meis@imbi.uni-heidelberg.de (ORCID)

Nico Bruder bruder@imbi.uni-heidelberg.de